+++ sdk/python/pulumi_kubernetes/provider.py
@@ -24,6 +24,8 @@ class Provider(pulumi.ProviderResource):
                                              If this is set, this config will be used instead of $KUBECONFIG.
         :param pulumi.Input[str] namespace: If present, the default namespace to use.
                                             This flag is ignored for cluster-scoped resources.
+                                            Note: if .metadata.namespace is set on a resource, that value takes
+                                            precedence over the provider default.
         """
         __props__ = {
             "cluster": cluster,