+++ sdk/python/pulumi_kubernetes/provider.py
@@ -20,9 +20,12 @@ class Provider(pulumi.ProviderResource):
         :param pulumi.ResourceOptions __opts__: An optional bag of options that controls this resource's behavior.
         :param pulumi.Input[str] cluster: If present, the name of the kubeconfig cluster to use.
         :param pulumi.Input[str] context: If present, the name of the kubeconfig context to use.
-        :param pulumi.Input[str] kubeconfig: The contents of a kubeconfig file. If this is set, this config will be used instead
-                               of $KUBECONFIG.
-        :param pulumi.Input[str] namespace: If present, the namespace scope to use.
+        :param pulumi.Input[str] kubeconfig: The contents of a kubeconfig file.
+                                             If this is set, this config will be used instead of $KUBECONFIG.
+        :param pulumi.Input[str] namespace: If present, the default namespace to use.
+                                            This flag is ignored for cluster-scoped resources.
+                                            Note: if .metadata.namespace is set on a resource, that value takes
+                                            precedence over the provider default.
         """
         __props__ = {
             "cluster": cluster,