+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 
 ### Improvements
 
--   None
+-   Enable configuring `ResourceOptions` via `transformations` (https://github.com/pulumi/pulumi-kubernetes/pull/575)
 
 ### Bug fixes