+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 ### Improvements
 
 -   Enable configuring `ResourceOptions` via `transformations` (https://github.com/pulumi/pulumi-kubernetes/pull/575).
--   Implement CheckConfig/DiffConfig (https://github.com/pulumi/pulumi-kubernetes/pull/577).
+-   Changing k8s cluster config now correctly causes dependent resources to be replaced (https://github.com/pulumi/pulumi-kubernetes/pull/577).
 
 ### Bug fixes